+++ src/app.module.ts
@@ -2,12 +2,12 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { GerenteModule } from './funcionarios/funcionario.module';
-import { ContaModule } from './contas/conta.module';
-import { ClienteModule } from './clientes/cliente.module';
-import { ClienteFabrica } from './clientes/fabricas/cliente.fabrica';
-import { ContaFabrica } from './contas/fabricas/conta.fabrica';
-import { FuncionarioFabrica } from './funcionarios/factories/funcionario.fabrica';
+import { ContaFabrica } from './domain/factories/conta.fabrica';
+import { FuncionarioFabrica } from './domain/factories/funcionario.fabrica';
+import { ClienteModule } from './modules/cliente.module';
+import { ContaModule } from './modules/conta.module';
+import { GerenteModule } from './modules/funcionario.module';
+import { ClienteFabrica } from './domain/factories/cliente.fabrica';
 
 @Module({
   imports: [ClienteModule, ContaModule, GerenteModule],
