--- src/clientes/tests/cliente.controller.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ClienteController } from './cliente.controller';
-
-describe('ClienteController', () => {
-  let controller: ClienteController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [ClienteController],
-    }).compile();
-
-    controller = module.get<ClienteController>(ClienteController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
