--- src/clientes/tests/cliente.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ClienteService } from './cliente.service';
-
-describe('ClienteService', () => {
-  let service: ClienteService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [ClienteService],
-    }).compile();
-
-    service = module.get<ClienteService>(ClienteService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
