+++ src/domain/enums/tipo-cliente.enum.ts
