+++ src/domain/enums/tipo-funcionario.enum.ts
