+++ src/domain/factories/funcionario.fabrica.ts
