+++ src/domain/interfaces/conta.interface.ts
@@ -1,0 +1,9 @@
+import { TipoCliente } from "src/domain/enums/tipo-cliente.enum"
+import { TipoConta } from "../enums/tipo-conta.enum"
+
+export interface Conta {
+    tipo: TipoConta
+    saldo: number
+    id: number,
+    clienteId: number
+}
