+++ src/domain/interfaces/funcionario.interface.ts
