+++ test/cliente.controller.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { ClienteController } from 'src/adapters/cliente.controller';
+
+describe('ClienteController', () => {
+  let controller: ClienteController;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [ClienteController],
+    }).compile();
+
+    controller = module.get<ClienteController>(ClienteController);
+  });
+
+  it('should be defined', () => {
+    expect(controller).toBeDefined();
+  });
+});
