+++ test/funcionario.controller.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { FuncionarioController } from 'src/adapters/funcionario.controller';
+
+describe('GerenteController', () => {
+  let controller: FuncionarioController;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [FuncionarioController],
+    }).compile();
+
+    controller = module.get<FuncionarioController>(FuncionarioController);
+  });
+
+  it('should be defined', () => {
+    expect(controller).toBeDefined();
+  });
+});
